+++ src/address/message.rs
@@ -1,10 +1,11 @@
 use std::future::Future;
-use std::marker::PhantomData;
 use std::pin::Pin;
 use std::task::{self, Poll};
 use std::time::Duration;
 
-use futures::channel::oneshot;
+use futures_channel::oneshot;
+
+use pin_project::pin_project;
 
 use crate::clock::Delay;
 use crate::handler::{Handler, Message};
@@ -15,6 +16,7 @@ use super::{MailboxError, SendError, ToEnvelope};
 /// A `Future` which represents an asynchronous message sending
 /// process.
 #[must_use = "You have to wait on request otherwise the Message wont be delivered"]
+#[pin_project]
 pub struct Request<A, M>
 where
     A: Handler<M>,
@@ -24,7 +26,6 @@ where
     rx: Option<oneshot::Receiver<M::Result>>,
     info: Option<(AddressSender<A>, M)>,
     timeout: Option<Delay>,
-    act: PhantomData<A>,
 }
 
 impl<A, M> Request<A, M>
@@ -41,7 +42,6 @@ where
             rx,
             info,
             timeout: None,
-            act: PhantomData,
         }
     }
 
@@ -61,7 +61,7 @@ where
         cx: &mut task::Context<'_>,
     ) -> Poll<Result<M::Result, MailboxError>> {
         if let Some(ref mut timeout) = self.timeout {
-            match unsafe { Pin::new_unchecked(timeout) }.poll(cx) {
+            match Pin::new(timeout).poll(cx) {
                 Poll::Ready(()) => Poll::Ready(Err(MailboxError::Timeout)),
                 Poll::Pending => Poll::Pending,
             }
@@ -80,14 +80,14 @@ where
 {
     type Output = Result<M::Result, MailboxError>;
 
-    fn poll(self: Pin<&mut Self>, cx: &mut task::Context<'_>) -> Poll<Self::Output> {
-        let mut this = unsafe { self.get_unchecked_mut() };
+    fn poll(mut self: Pin<&mut Self>, cx: &mut task::Context<'_>) -> Poll<Self::Output> {
+        let this = self.as_mut().project();
 
         if let Some((sender, msg)) = this.info.take() {
             match sender.send(msg) {
-                Ok(rx) => this.rx = Some(rx),
+                Ok(rx) => *this.rx = Some(rx),
                 Err(SendError::Full(msg)) => {
-                    this.info = Some((sender, msg));
+                    *this.info = Some((sender, msg));
                     return Poll::Pending;
                 }
                 Err(SendError::Closed(_)) => {
@@ -97,11 +97,10 @@ where
         }
 
         if this.rx.is_some() {
-            match unsafe { Pin::new_unchecked(&mut this.rx.as_mut().unwrap()) }.poll(cx)
-            {
+            match Pin::new(&mut this.rx.as_mut().unwrap()).poll(cx) {
                 Poll::Ready(Ok(i)) => Poll::Ready(Ok(i)),
                 Poll::Ready(Err(_)) => Poll::Ready(Err(MailboxError::Closed)),
-                Poll::Pending => this.poll_timeout(cx),
+                Poll::Pending => self.poll_timeout(cx),
             }
         } else {
             Poll::Ready(Err(MailboxError::Closed))
@@ -111,6 +110,7 @@ where
 
 /// A `Future` which represents an asynchronous message sending process.
 #[must_use = "future do nothing unless polled"]
+#[pin_project]
 pub struct RecipientRequest<M>
 where
     M: Message + Send + 'static,
@@ -148,7 +148,7 @@ where
         cx: &mut task::Context<'_>,
     ) -> Poll<Result<M::Result, MailboxError>> {
         if let Some(ref mut timeout) = self.timeout {
-            match unsafe { Pin::new_unchecked(timeout) }.poll(cx) {
+            match Pin::new(timeout).poll(cx) {
                 Poll::Ready(()) => Poll::Ready(Err(MailboxError::Timeout)),
                 Poll::Pending => Poll::Pending,
             }
@@ -165,14 +165,14 @@ where
 {
     type Output = Result<M::Result, MailboxError>;
 
-    fn poll(self: Pin<&mut Self>, cx: &mut task::Context<'_>) -> Poll<Self::Output> {
-        let mut this = unsafe { self.get_unchecked_mut() };
+    fn poll(mut self: Pin<&mut Self>, cx: &mut task::Context<'_>) -> Poll<Self::Output> {
+        let this = self.as_mut().project();
 
         if let Some((sender, msg)) = this.info.take() {
             match sender.send(msg) {
-                Ok(rx) => this.rx = Some(rx),
+                Ok(rx) => *this.rx = Some(rx),
                 Err(SendError::Full(msg)) => {
-                    this.info = Some((sender, msg));
+                    *this.info = Some((sender, msg));
                     return Poll::Pending;
                 }
                 Err(SendError::Closed(_)) => {
@@ -182,11 +182,10 @@ where
         }
 
         if this.rx.is_some() {
-            match unsafe { Pin::new_unchecked(&mut this.rx.as_mut().unwrap()) }.poll(cx)
-            {
+            match Pin::new(&mut this.rx.as_mut().unwrap()).poll(cx) {
                 Poll::Ready(Ok(i)) => Poll::Ready(Ok(i)),
                 Poll::Ready(Err(_)) => Poll::Ready(Err(MailboxError::Closed)),
-                Poll::Pending => this.poll_timeout(cx),
+                Poll::Pending => self.poll_timeout(cx),
             }
         } else {
             Poll::Ready(Err(MailboxError::Closed))