+++ src/fut/mod.rs
@@ -4,7 +4,7 @@ use std::marker::PhantomData;
 use std::task::{Context, Poll};
 use std::time::Duration;
 
-use futures::{Future, Stream};
+use futures_util::{future::Future, stream::Stream};
 use pin_project::pin_project;
 
 mod chain;
@@ -122,7 +122,7 @@ use std::pin::Pin;
 ///         });
 ///
 ///         // return the wrapped future
-///         Box::new(update_self)
+///         Box::pin(update_self)
 ///     }
 /// }
 ///
@@ -186,7 +186,7 @@ pub trait ActorFuture {
 
 /// A stream of values, not all of which may have been produced yet.
 ///
-/// This is similar to `futures::Stream` trait, except it works with `Actor`
+/// This is similar to `futures_util::stream::Stream` trait, except it works with `Actor`
 pub trait ActorStream {
     /// The type of item this stream will yield on success.
     type Item;
@@ -292,7 +292,7 @@ impl<F: ActorFuture> IntoActorFuture for F {
     }
 }
 
-impl<F: ActorFuture + ?Sized> ActorFuture for Box<F> {
+impl<F: ActorFuture + Unpin + ?Sized> ActorFuture for Box<F> {
     type Output = F::Output;
     type Actor = F::Actor;
 
@@ -302,7 +302,25 @@ impl<F: ActorFuture + ?Sized> ActorFuture for Box<F> {
         ctx: &mut <Self::Actor as Actor>::Context,
         task: &mut Context<'_>,
     ) -> Poll<Self::Output> {
-        unsafe { Pin::new_unchecked(&mut **self.as_mut()) }.poll(srv, ctx, task)
+        Pin::new(&mut **self.as_mut()).poll(srv, ctx, task)
+    }
+}
+
+impl<P> ActorFuture for Pin<P>
+where
+    P: Unpin + std::ops::DerefMut,
+    <P as std::ops::Deref>::Target: ActorFuture,
+{
+    type Output = <<P as std::ops::Deref>::Target as ActorFuture>::Output;
+    type Actor = <<P as std::ops::Deref>::Target as ActorFuture>::Actor;
+
+    fn poll(
+        self: Pin<&mut Self>,
+        srv: &mut Self::Actor,
+        ctx: &mut <Self::Actor as Actor>::Context,
+        task: &mut Context<'_>,
+    ) -> Poll<Self::Output> {
+        Pin::get_mut(self).as_mut().poll(srv, ctx, task)
     }
 }